lazy.c: fix distinct pointer types lacks a cast

+++ drivers/cpufreq/cpufreq_lazy.c
@@ -733,7 +733,7 @@ static int cpufreq_governor_dbs(struct cpufreq_policy *policy,
 	    /* Bring kernel and HW constraints together */
 	    min_sampling_rate = max(min_sampling_rate,
 				    MIN_LATENCY_MULTIPLIER * latency);
-	    dbs_tuners_ins.sampling_rate = max(min_sampling_rate, DEF_SAMPLE_RATE);
+	    dbs_tuners_ins.sampling_rate = max((int)min_sampling_rate, DEF_SAMPLE_RATE);
 	    current_sampling_rate = dbs_tuners_ins.sampling_rate;
 	    dbs_tuners_ins.min_timeinstate = latency * LATENCY_MULTIPLIER;
 	    dbs_tuners_ins.min_timeinstate = 4 * max(dbs_tuners_ins.sampling_rate, dbs_tuners_ins.min_timeinstate);
